backend/requirements.txt
flask>=3.0
flask-cors>=4.0
tensorflow>=2.15
Pillow>=10.0
numpy>=1.24
